ai/ai-dataset/requirements.txt
openai>=1.40
pydantic>=2.0
PyYAML>=6.0
python-dotenv>=1.0
scikit-learn>=1.3
numpy>=1.24
